Shiny fixes (#803) (#804)
* Conversion to version 3

* Add export Cohort Definitions
* Characterization changes includes optimization, efficient use of reactive objects
* Updated timeDistribution, characterization, overlap. Time distribution is now part of characterization output. Overlap is part of cohortRelationship
* Code formatted using styler
* lot of bug fixes
* Skip testing annotation table for data removal
* Fix to markdown - no longer save to tempfile just uses markdown package
* Added zip utility for creating sharable diagnostics explorers

+++ tests/testthat/test-ResultsDataModel.R
@@ -234,36 +234,44 @@ test_that("Sqlite results data model", {
 
 
 test_that("Data removal works", {
-  skip_if(skipResultsDm | skipCdmTests, "results data model test server not set")
+  skip_if(skipResultsDm |
+            skipCdmTests,
+          "results data model test server not set")
   specifications <- getResultsDataModelSpecifications()
-
-  pgConnection <- DatabaseConnector::connect(connectionDetails = postgresConnectionDetails)
+  
+  pgConnection <-
+    DatabaseConnector::connect(connectionDetails = postgresConnectionDetails)
   with_dbc_connection(pgConnection, {
     for (tableName in unique(specifications$tableName)) {
-      primaryKey <- specifications %>%
-        dplyr::filter(.data$tableName == !!tableName &
-          .data$primaryKey == "Yes") %>%
-        dplyr::select(.data$fieldName) %>%
-        dplyr::pull()
-
-      if ("database_id" %in% primaryKey) {
-        deleteAllRecordsForDatabaseId(
-          connection = pgConnection,
-          schema = resultsDatabaseSchema,
-          tableName = tableName,
-          databaseId = "cdmv5"
-        )
-
-        sql <- "SELECT COUNT(*) FROM @schema.@table_name WHERE database_id = '@database_id';"
-        sql <- SqlRender::render(
-          sql = sql,
-          schema = resultsDatabaseSchema,
-          table_name = tableName,
-          database_id = "cdmv5"
-        )
-        databaseIdCount <-
-          DatabaseConnector::querySql(pgConnection, sql)[, 1]
-        expect_true(databaseIdCount == 0)
+      if (stringr::str_detect(string = tolower(tableName),
+                              pattern = "annotation",
+                              negate = TRUE)) {
+        primaryKey <- specifications %>%
+          dplyr::filter(.data$tableName == !!tableName &
+                          .data$primaryKey == "Yes") %>%
+          dplyr::select(.data$fieldName) %>%
+          dplyr::pull()
+        
+        if ("database_id" %in% primaryKey) {
+          deleteAllRecordsForDatabaseId(
+            connection = pgConnection,
+            schema = resultsDatabaseSchema,
+            tableName = tableName,
+            databaseId = "cdmv5"
+          )
+          
+          sql <-
+            "SELECT COUNT(*) FROM @schema.@table_name WHERE database_id = '@database_id';"
+          sql <- SqlRender::render(
+            sql = sql,
+            schema = resultsDatabaseSchema,
+            table_name = tableName,
+            database_id = "cdmv5"
+          )
+          databaseIdCount <-
+            DatabaseConnector::querySql(pgConnection, sql)[, 1]
+          expect_true(databaseIdCount == 0)
+        }
       }
     }
   })
